+++ app/src/main/java/com/github/bumblebee202111/doubean/data/repository/SearchSubjectsRepository.kt
@@ -1,11 +1,12 @@
 package com.github.bumblebee202111.doubean.data.repository
 
-import com.github.bumblebee202111.doubean.coroutines.suspendRunCatching
-import com.github.bumblebee202111.doubean.model.SearchResultSubjectItem
+import com.github.bumblebee202111.doubean.model.AppResult
+import com.github.bumblebee202111.doubean.model.SubjectSearchResult
 import com.github.bumblebee202111.doubean.model.SubjectsSearchType
+import com.github.bumblebee202111.doubean.model.safeApiCall
 import com.github.bumblebee202111.doubean.model.toApiSubjectsSearchType
 import com.github.bumblebee202111.doubean.network.ApiService
-import com.github.bumblebee202111.doubean.network.model.toSearchResultSubjectItems
+import com.github.bumblebee202111.doubean.network.model.toSubjectSearchResult
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -13,10 +14,16 @@ import javax.inject.Singleton
 class SearchSubjectsRepository @Inject constructor(private val apiService: ApiService) {
     suspend fun searchSubjects(
         query: String,
-        type: SubjectsSearchType,
-    ): Result<List<SearchResultSubjectItem>> {
-        return suspendRunCatching {
-            apiService.searchSubjects(query, type.toApiSubjectsSearchType())
-        }.map { it.subjects.toSearchResultSubjectItems() }
+        type: SubjectsSearchType? = null,
+    ): AppResult<SubjectSearchResult> {
+        return safeApiCall(
+            apiCall = {
+                apiService.searchSubjects(q = query, type = type?.toApiSubjectsSearchType())
+            },
+            mapSuccess = {
+                it.toSubjectSearchResult()
+            }
+        )
     }
+
 }