+++ app/src/main/java/com/github/bumblebee202111/doubean/navigation/BottomNavHost.kt
@@ -9,7 +9,6 @@ import com.github.bumblebee202111.doubean.feature.profile.navigation.profileScre
 import com.github.bumblebee202111.doubean.feature.statuses.navigation.statusesScreen
 import com.github.bumblebee202111.doubean.feature.subjects.navigation.subjectsScreen
 import com.github.bumblebee202111.doubean.model.SubjectType
-import com.github.bumblebee202111.doubean.model.SubjectsSearchType
 import com.github.bumblebee202111.doubean.ui.BottomNavDestination
 
 @Composable
@@ -23,7 +22,7 @@ fun BottomNavHost(
     navigateToTopic: (topicId: String) -> Unit,
     navigateToLogin: () -> Unit,
     navigateToSubjectInterests: (userId: String, subjectType: SubjectType) -> Unit,
-    navigateToSearchSubjects: (type: SubjectsSearchType) -> Unit,
+    navigateToSearchSubjects: () -> Unit,
     navigateToRankList: (collectionId: String) -> Unit,
     navigateToMovie: (movieId: String) -> Unit,
     navigateToTv: (tvId: String) -> Unit,